+++ packages/neuron-ui/src/components/AmendSUDTSend/index.tsx
@@ -18,6 +18,7 @@ import {
 } from 'utils'
 import { DEFAULT_SUDT_FIELDS } from 'utils/const'
 import AlertDialog from 'widgets/AlertDialog'
+import { useOutputs } from 'components/AmendPendingTransactionDialog/hooks'
 import styles from './amendSUDTSend.module.scss'
 import { useInitialize } from './hooks'
 
@@ -48,7 +49,7 @@ const AmendSUDTSend = () => {
     transaction,
     onSubmit,
     minPrice,
-    showConfirmedAlert,
+    isConfirmedAlertShown,
     sudtInfo,
     description,
     onDescriptionChange,
@@ -61,6 +62,14 @@ const AmendSUDTSend = () => {
     t,
   })
 
+  const { items, lastOutputsCapacity } = useOutputs({
+    transaction,
+    isMainnet,
+    addresses,
+    sUDTAccounts,
+    fee,
+  })
+
   const priceError = useMemo(() => {
     return Number(price || '0') < Number(minPrice) ? t('price-switch.errorTip', { minPrice }) : null
   }, [price, minPrice])
@@ -71,7 +80,7 @@ const AmendSUDTSend = () => {
     (e: React.SyntheticEvent<HTMLInputElement>) => {
       const { value: inputValue } = e.currentTarget
 
-      const value = inputValue.split('.')[0].replace(/[^\d]/, '')
+      const value = inputValue.split('.')[0].replace(/[^\d]/g, '')
       setPrice(value)
     },
     [setPrice]
@@ -95,83 +104,8 @@ const AmendSUDTSend = () => {
     return scriptToAddress(transaction?.outputs[0].lock, { isMainnet })
   }, [transaction?.outputs])
 
-  const getLastOutputAddress = (outputs: State.DetailedOutput[]) => {
-    const change = outputs.find(output => {
-      const address = scriptToAddress(output.lock, { isMainnet })
-
-      return !!addresses.find(item => item.address === address && item.type === 1)
-    })
-    if (change) {
-      return scriptToAddress(change.lock, { isMainnet })
-    }
-
-    const receive = outputs.find(output => {
-      const address = scriptToAddress(output.lock, { isMainnet })
-      return !!addresses.find(item => item.address === address && item.type === 0)
-    })
-    if (receive) {
-      return scriptToAddress(receive.lock, { isMainnet })
-    }
-
-    const sudt = outputs.find(output => {
-      const address = scriptToAddress(output.lock, { isMainnet })
-      return !!sUDTAccounts.find(item => item.address === address)
-    })
-    if (sudt) {
-      return scriptToAddress(sudt.lock, { isMainnet })
-    }
-    return ''
-  }
-
-  const items: {
-    address: string
-    amount: string
-    capacity: string
-    isLastOutput: boolean
-    output: State.DetailedOutput
-  }[] = useMemo(() => {
-    if (transaction && transaction.outputs.length) {
-      const lastOutputAddress = getLastOutputAddress(transaction.outputs)
-      return transaction.outputs.map(output => {
-        const address = scriptToAddress(output.lock, { isMainnet })
-        return {
-          capacity: output.capacity,
-          address,
-          output,
-          amount: shannonToCKBFormatter(output.capacity || '0'),
-          isLastOutput: address === lastOutputAddress,
-        }
-      })
-    }
-    return []
-  }, [transaction?.outputs])
-
-  const outputsCapacity = useMemo(() => {
-    const outputList = items.filter(item => !item.isLastOutput)
-    return outputList.reduce((total, cur) => {
-      if (Number.isNaN(+(cur.capacity || ''))) {
-        return total
-      }
-      return total + BigInt(cur.capacity || '0')
-    }, BigInt(0))
-  }, [items])
-
-  const lastOutputsCapacity = useMemo(() => {
-    if (transaction) {
-      const inputsCapacity = transaction.inputs.reduce((total, cur) => {
-        if (Number.isNaN(+(cur.capacity || ''))) {
-          return total
-        }
-        return total + BigInt(cur.capacity || '0')
-      }, BigInt(0))
-
-      return inputsCapacity - outputsCapacity - fee
-    }
-    return -1
-  }, [transaction, fee, outputsCapacity])
-
   useEffect(() => {
-    if (transaction) {
+    if (transaction && lastOutputsCapacity !== undefined) {
       const outputs = items.map(item => {
         const capacity = item.isLastOutput ? lastOutputsCapacity.toString() : item.capacity
         return {
@@ -192,7 +126,8 @@ const AmendSUDTSend = () => {
     }
   }, [lastOutputsCapacity, transaction, items, dispatch, experimental?.params?.description, description])
 
-  const disabled = sending || !experimental?.tx || priceError || lastOutputsCapacity < 0
+  const disabled =
+    sending || !experimental?.tx || priceError || lastOutputsCapacity === undefined || lastOutputsCapacity < 0
 
   return (
     <PageContainer
@@ -263,7 +198,7 @@ const AmendSUDTSend = () => {
             </div>
             <div className={styles.rightFooter}>
               <div className={styles.actions}>
-                <Button type="submit" disabled={disabled} label={t('send.send')}>
+                <Button type="submit" disabled={!!disabled} label={t('send.send')}>
                   {sending ? <Spinner /> : (t('send.submit-transaction') as string)}
                 </Button>
               </div>
@@ -272,7 +207,7 @@ const AmendSUDTSend = () => {
         </div>
       </form>
       <AlertDialog
-        show={showConfirmedAlert}
+        show={isConfirmedAlertShown}
         title={t('send.transaction-confirmed')}
         message={t('send.transaction-cannot-amend')}
         type="warning"