+++ packages/neuron-wallet/src/controllers/api.ts
@@ -59,7 +59,7 @@ import startMonitor, { stopMonitor } from '../services/monitor'
 import { migrateCkbData } from '../services/ckb-runner'
 import NodeService from '../services/node'
 import SyncProgressService from '../services/sync-progress'
-import { resetSyncTaskQueue } from '../block-sync-renderer'
+import { changeMultisigSyncStatus, resetSyncTaskQueue } from '../block-sync-renderer'
 import DataUpdateSubject from '../models/subjects/data-update'
 import CellManagement from './cell-management'
 import { UpdateCellLocalInfo } from '../database/chain/entities/cell-local-info'
@@ -449,6 +449,7 @@ export default class ApiController {
           description?: string
           multisigConfig?: MultisigConfigModel
           amendHash?: string
+          skipLastInputs?: boolean
         }
       ) => {
         return this.#walletsController.sendTx({
@@ -835,10 +836,6 @@ export default class ApiController {
       return this.#hardwareController.getCkbAppVersion()
     })
 
-    handle('get-device-firmware-version', async () => {
-      return this.#hardwareController.getFirmwareVersion()
-    })
-
     handle('get-device-extended-public-key', async () => {
       return this.#hardwareController.getExtendedPublicKey()
     })
@@ -866,10 +863,6 @@ export default class ApiController {
 
     handle('broadcast-signed-transaction', async (_, params) => {
       return this.#offlineSignController.broadcastTransaction({ ...params, walletID: '' })
-    })
-
-    handle('get-transaction-size', async (_, params) => {
-      return this.#transactionsController.getTransactionSize(params)
     })
 
     handle('sign-and-export-transaction', async (_, params) => {
@@ -918,6 +911,13 @@ export default class ApiController {
 
     handle('load-multisig-tx-json', async (_, fullPayload) => {
       return this.#multisigController.loadMultisigTxJson(fullPayload)
+    })
+
+    handle('change-multisig-sync-status', (_, status: boolean) => {
+      changeMultisigSyncStatus(status)
+      return {
+        status: ResponseCode.Success,
+      }
     })
 
     //migrate