+++ packages/neuron-wallet/src/services/ckb-runner.ts
@@ -91,10 +91,6 @@ const initCkb = async () => {
   })
 }
 
-let isLookingValidTarget: boolean = false
-let lastLogTime: number
-export const getLookingValidTargetStatus = () => isLookingValidTarget
-
 export const getNodeUrl = () => `${BUNDLED_URL_PREFIX}${rpcPort}`
 
 const removeOldIndexerIfRunSuccess = () => {
@@ -115,8 +111,12 @@ export const startCkbNode = async () => {
   listenPort = await getUsablePort(rpcPort >= listenPort ? rpcPort + 1 : listenPort)
 
   updateToml(path.join(SettingsService.getInstance().getNodeDataPath(), 'ckb.toml'), {
-    rpc: `listen_address = "127.0.0.1:${rpcPort}"`,
-    network: `listen_addresses = ["/ip4/0.0.0.0/tcp/${listenPort}"]`,
+    rpc: {
+      listen_address: `"127.0.0.1:${rpcPort}"`,
+    },
+    network: {
+      listen_addresses: `["/ip4/0.0.0.0/tcp/${listenPort}"]`,
+    },
   })
   const options = ['run', '-C', SettingsService.getInstance().getNodeDataPath(), '--indexer']
   const stdio: (StdioNull | StdioPipe)[] = ['ignore', 'pipe', 'pipe']
@@ -140,21 +140,10 @@ export const startCkbNode = async () => {
       logger.error('CKB:\trun fail:', dataString)
       return
     }
-    if (
-      dataString.includes(
-        `can't find assume valid target temporarily, hash: Byte32(${process.env.CKB_NODE_ASSUME_VALID_TARGET})`
-      )
-    ) {
-      isLookingValidTarget = true
-      lastLogTime = Date.now()
-    } else if (lastLogTime && Date.now() - lastLogTime > 10000) {
-      isLookingValidTarget = false
-    }
   })
 
   currentProcess.on('error', error => {
     logger.error('CKB:\trun fail:', error)
-    isLookingValidTarget = false
     if (Object.is(ckb, currentProcess)) {
       ckb = null
     }
@@ -162,7 +151,6 @@ export const startCkbNode = async () => {
 
   currentProcess.on('close', code => {
     logger.info(`CKB:\tprocess closed with code ${code}`)
-    isLookingValidTarget = false
     if (Object.is(ckb, currentProcess)) {
       ckb = null
     }